Issue #752 Browse BAM coverage intervals

+++ server/catgenome/src/main/java/com/epam/catgenome/entity/bam/CoverageInterval.java
@@ -35,5 +35,5 @@ public class CoverageInterval {
     private String chr;
     private Integer start;
     private Integer end;
-    private Integer coverage;
+    private Float coverage;
 }
